pyproject.toml
[project]
name = "mystique-ai"
version = "0.1.0"
description = "A shapeshifting LLM wrapper with multi-provider support."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.32.0",
    "litellm>=1.34.0",
    "tinydb>=4.8.0",
    "pyyaml>=6.0.1",
    "pandas>=2.2.0",
    "plotly>=5.19.0",
    "watchdog>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# --- FIX 1: Explicitly map the package ---
[tool.hatch.build.targets.wheel]
packages = ["src/mystique"]

# --- FIX 2: Use the new standard for dev dependencies ---
[dependency-groups]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true